requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.6
autopep8==1.5.7
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
cryptography==3.3.1
cssselect==1.1.0
defusedxml==0.6.0
Django==3.2.3
django-admin-list-filter-dropdown==1.0.3
django-allauth==0.44.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-suit==0.2.28
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
fake-useragent==0.1.11
flake8==3.8.3
idna==2.10
importlib-metadata==2.0.0
isort==5.8.0
jedi==0.17.2
lazy-object-proxy==1.6.0
lxml==4.5.2
Markdown==3.3.4
mccabe==0.6.1
oauthlib==3.1.0
parse==1.18.0
parso==0.7.1
Pillow==8.2.0
pluggy==0.13.1
pycodestyle==2.7.0
pycparser==2.20
pydocstyle==5.1.1
pyee==7.0.4
pyflakes==2.2.0
PyJWT==1.7.1
pylama==7.7.1
pylama-pylint==3.1.1
pylint==2.8.2
pyppeteer==0.2.2
pyquery==1.4.1
python-jsonrpc-server==0.4.0
python-language-server==0.35.1
python3-openid==3.2.0
pytz==2021.1
requests==2.24.0
requests-html==0.10.0
requests-oauthlib==1.3.0
rope==0.17.0
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
tqdm==4.49.0
typed-ast==1.4.1
typing-extensions==3.10.0.0
ujson==3.2.0
urllib3==1.25.10
w3lib==1.22.0
websockets==8.1
wincertstore==0.2
wrapt==1.12.1
yapf==0.30.0
zipp==3.2.0
